+++ changelog.md
@@ -1,3 +1,7 @@
+## 1.1.0
+- Added Levels compatability.
+- Made some changes to increase stability and maybe performance.
+
 ## 1.0.2
 - updated manifest to include bugs and readme